requirements.txt

Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Flask-Cors==3.0.10
Werkzeug==2.3.0